requirements.txt

pandas
numpy
torch           
jupyterlab
ipywidgets
sentencepiece

ollama
faiss-cpu
langchain
transformers
sentence-transformers
accelerate